create common module and move ActiveSpan and RumConstants.
