+++ common/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+    id("otel.android-library-conventions")
+    id("otel.publish-conventions")
+}
+
+description = "OpenTelemetry Android common"
+
+android {
+    namespace = "io.opentelemetry.android.common"
+
+    defaultConfig {
+        consumerProguardFiles("consumer-rules.pro")
+    }
+}
+
+dependencies {
+    api(platform(libs.opentelemetry.platform))
+    api(libs.opentelemetry.api)
+    implementation(libs.androidx.core)
+}
