+++ instrumentation/src/main/java/io/opentelemetry/android/SessionIdSpanAppender.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.android;
 
+import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.trace.ReadWriteSpan;
 import io.opentelemetry.sdk.trace.ReadableSpan;
