+++ instrumentation/src/main/java/io/opentelemetry/android/instrumentation/slowrendering/SlowRenderListener.java
@@ -21,7 +21,7 @@ import androidx.annotation.GuardedBy;
 import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
 import io.opentelemetry.android.DefaultingActivityLifecycleCallbacks;
-import io.opentelemetry.android.RumConstants;
+import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.Tracer;
 import java.time.Duration;
