+++ instrumentation/src/main/java/io/opentelemetry/android/instrumentation/slowrendering/SlowRenderingDetector.java
@@ -7,7 +7,7 @@ package io.opentelemetry.android.instrumentation.slowrendering;
 
 import android.os.Build;
 import android.util.Log;
-import io.opentelemetry.android.RumConstants;
+import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.android.instrumentation.InstrumentedApplication;
 import java.time.Duration;
