+++ instrumentation/src/main/java/io/opentelemetry/android/instrumentation/startup/AppStartupTimer.java
@@ -13,7 +13,7 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import io.opentelemetry.android.DefaultingActivityLifecycleCallbacks;
-import io.opentelemetry.android.RumConstants;
+import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.android.util.AnchoredClock;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.Tracer;
