+++ instrumentation/src/test/java/io/opentelemetry/android/instrumentation/activity/ActivityTracerTest.java
@@ -15,7 +15,7 @@ import static org.mockito.Mockito.when;
 
 import android.app.Activity;
 import io.opentelemetry.android.instrumentation.startup.AppStartupTimer;
-import io.opentelemetry.android.util.ActiveSpan;
+import io.opentelemetry.android.common.ActiveSpan;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension;
 import io.opentelemetry.sdk.trace.data.SpanData;
