+++ instrumentation/src/test/java/io/opentelemetry/android/instrumentation/fragment/FragmentTracerTest.java
@@ -13,7 +13,7 @@ import static org.mockito.Mockito.when;
 
 import androidx.fragment.app.Fragment;
 import io.opentelemetry.android.instrumentation.activity.VisibleScreenTracker;
-import io.opentelemetry.android.util.ActiveSpan;
+import io.opentelemetry.android.common.ActiveSpan;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension;
 import io.opentelemetry.sdk.trace.data.SpanData;
