+++ settings.gradle.kts
@@ -1,5 +1,6 @@
 rootProject.name = "opentelemetry-android"
 
+include(":common")
 include(":instrumentation")
 include(":auto-instrumentation:okhttp:okhttp-3.0:agent")
 include(":auto-instrumentation:okhttp:okhttp-3.0:library")
